make hazard detection and forwarding use signals
Previously, these units signaled pipelined registers to freeze or
flush by directly setting their member variables such as
`shouldFlush` or `shouldFreeze`. This was not a problem when we
were only operating on the issue unit and the IF-ID pipeline
registers but with the possible introduction of a timed memory
system, these restrictions need to be lifted and the mechanisms
generalized, hence the move to signals.